refactor(discover-page): rename discover.* to discover-page.*
Refs #65
