+++ apps/mull-ui/src/app/app.tsx
@@ -14,7 +14,7 @@ import { ROUTES } from '../constants';
 import 'react-toastify/dist/ReactToastify.min.css';
 import './app.scss';
 import { dummyEvent } from '../constants';
-import DiscoverPage from './pages/home/discover/discover';
+import DiscoverPage from './pages/home/discover/discover-page';
 import UpcomingPage from './pages/home/upcoming/upcoming';
 import MyEventsPage from './pages/home/my-events/my-events';
 import EventPage from './pages/event-page/event-page';
